Fix: moved to autocommands
