+++ src/main/java/frc/robot/RobotContainer.java
@@ -4,7 +4,7 @@ import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.RobotController;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
-import frc.robot.utilities.Questionnaire;
+import frc.robot.commands.autocommands.Questionnaire;
 import frc.lib.util.flippable.Flippable;
 import frc.robot.poseestimation.poseestimator.PoseEstimator;
 import frc.robot.subsystems.algaeblaster.AlgaeBlaster;
@@ -31,8 +31,7 @@ public class RobotContainer {
     public static final AlgaeBlaster ALGAE_BLASTER = new AlgaeBlaster();
     public static final AlgaeIntake ALGAE_INTAKE = new AlgaeIntake();
     public static final Leds LEDS = new Leds();
-
-    public static Questionnaire questionnaire;
+    public static final Questionnaire QUESTIONNAIRE = new Questionnaire();
 
     public RobotContainer() {
         DriverStation.silenceJoystickConnectionWarning(true);
@@ -40,18 +39,17 @@ public class RobotContainer {
         Flippable.init();
         PathPlannerConstants.initializePathPlanner();
 
-        setupAutonomous();
         setupLEDs();
 
         ButtonControls.initializeButtons(ButtonControls.ButtonLayout.TELEOP);
     }
 
     public Command getAutonomousCommand() {
-        return questionnaire.getCommand();
+        return QUESTIONNAIRE.getCommand();
     }
 
     public String getAutoName() {
-        return questionnaire.getSelected();
+        return QUESTIONNAIRE.getSelected();
     }
 
     private void setupLEDs() {
@@ -68,9 +66,5 @@ public class RobotContainer {
         });
 
         batteryLowTrigger.onTrue(LEDS.setLEDStatus(Leds.LEDMode.BATTERY_LOW, 5));
-    }
-
-    private void setupAutonomous() {
-        questionnaire = new Questionnaire();
     }
 }
